separate playwright and vitest code
