+++ deployables/app/e2e/accountHandling/unavailableWallet.spec.ts
@@ -1,6 +1,6 @@
 import test, { expect } from '@playwright/test'
 import { encode } from '@zodiac/schema'
-import { mockWeb3 } from '@zodiac/test-utils/e2e'
+import { mockWeb3 } from '@zodiac/test-utils/web3'
 import { href } from 'react-router'
 import { connectWallet } from '../connectWallet'
 import { account, route, transaction } from './fixture'
