+++ deployables/app/e2e/accountHandling/wrongChain.spec.ts
@@ -1,7 +1,7 @@
 import test, { expect } from '@playwright/test'
 import { Chain } from '@zodiac/chains'
 import { encode } from '@zodiac/schema'
-import { mockWeb3 } from '@zodiac/test-utils/e2e'
+import { mockWeb3 } from '@zodiac/test-utils/web3'
 import { href } from 'react-router'
 import { connectWallet } from '../connectWallet'
 import { account, route, transaction } from './fixture'
