+++ packages/test-utils/package.json
@@ -26,6 +26,10 @@
     "./e2e": {
       "types": "./src/e2e/index.ts",
       "import": "./src/e2e/index.ts"
+    },
+    "./web3": {
+      "types": "./src/web3/index.ts",
+      "import": "./src/web3/index.ts"
     }
   },
   "dependencies": {
