+++ packages/test-utils/src/e2e/index.ts
@@ -1,2 +1,1 @@
-export { defaultMockAccount, mockWeb3 } from './mockWeb3'
 export { waitFor } from './waitFor'
