+++ packages/test-utils/src/web3/index.ts
@@ -1,1 +1,2 @@
 export { MockJsonRpcProvider } from './MockJsonRpcProvider'
+export { defaultMockAccount, mockWeb3 } from './mockWeb3'
